requirements.txt
streamlit
streamlit-chat
langchain-community
langchain-core
python-dotenv
langchain-huggingface